refactor: 클린 아키텍쳐 리팩토링 (이게 맞는지는 모르겠음 그냥 실험) #135
