+++ be15_DevEagles_BE/src/main/java/com/deveagles/be15_deveagles_be/features/coupons/domain/service/CouponDiscountCalculator.java
@@ -1,0 +1,42 @@
+package com.deveagles.be15_deveagles_be.features.coupons.domain.service;
+
+import com.deveagles.be15_deveagles_be.features.coupons.common.CouponResponseFactory;
+import com.deveagles.be15_deveagles_be.features.coupons.domain.entity.Coupon;
+import com.deveagles.be15_deveagles_be.features.coupons.presentation.dto.response.CouponApplicationResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class CouponDiscountCalculator {
+
+  private final CouponResponseFactory couponResponseFactory;
+
+  public CouponApplicationResponse calculateDiscount(Coupon coupon, Integer originalAmount) {
+    log.info(
+        "할인 계산 시작 - 쿠폰ID: {}, 원금액: {}, 할인율: {}%",
+        coupon.getId(), originalAmount, coupon.getDiscountRate());
+
+    if (originalAmount == null || originalAmount <= 0) {
+      return couponResponseFactory.createFailedResponse("유효하지 않은 금액입니다");
+    }
+
+    Integer discountRate = coupon.getDiscountRate();
+    Integer discountAmount = calculateDiscountAmount(originalAmount, discountRate);
+    Integer finalAmount = originalAmount - discountAmount;
+
+    log.info("할인 계산 완료 - 할인율: {}%, 할인금액: {}, 최종금액: {}", discountRate, discountAmount, finalAmount);
+
+    return couponResponseFactory.createSuccessResponse(
+        coupon, discountRate, discountAmount, finalAmount);
+  }
+
+  private Integer calculateDiscountAmount(Integer originalAmount, Integer discountRate) {
+    if (originalAmount == null || discountRate == null) {
+      return 0;
+    }
+    return originalAmount * discountRate / 100;
+  }
+}
