+++ be15_DevEagles_BE/src/main/java/com/deveagles/be15_deveagles_be/features/coupons/infrastructure/repository/CouponRepository.java
@@ -1,0 +1,19 @@
+package com.deveagles.be15_deveagles_be.features.coupons.infrastructure.repository;
+
+import com.deveagles.be15_deveagles_be.features.coupons.domain.entity.Coupon;
+import com.deveagles.be15_deveagles_be.features.coupons.domain.repository.CouponQueryRepository;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CouponRepository extends JpaRepository<Coupon, Long>, CouponQueryRepository {
+
+  Optional<Coupon> findByCouponCodeAndDeletedAtIsNull(String couponCode);
+
+  @Query(
+      "SELECT COUNT(c) > 0 FROM Coupon c WHERE c.couponCode = :couponCode AND c.deletedAt IS NULL")
+  boolean existsByCouponCodeAndNotDeleted(@Param("couponCode") String couponCode);
+}
